fix props and linters error
