--- src/component/Button.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function Button(props) {
-return (
-<button type="button" className={props.className} onClick={props.onClick}>
-{props.label}
-</button>
-);
-}
-
-export default Button;
-
